refactor(state): extract deck store factory
